It works! (BUILD FIXED)
It took me a while to get a hang of it.

Okay,

rc.bat

WITHIN the DOSBOX (or equivalent) environment
will build KC, and move KC.EXE to the BUILT
sub-directory.

CD there and run the exe,
or run one of the older ones: KC1.EXE or KG.EXE

I'm adding them to the source control because, I hear,
Git works fine for "binary" (non-text) files, too also,
and I would like people to not have to build any of
my projects from source.

Those ways are dark wizardry.

In case it is not clear, run ("involk") rc.bat from
the root of the project directory.

Now, it needs DJGPP bash installed;
not sure if it would work with, say, Cygwin or whatever,
but that's academic at this point.

The source is horribe because it's like over 15 years olde,
and in "Internet time" that's about a century. Since then,
I learned how to refactor so well that I would never design
this project this way or release it in this state.

...but it works and looks pretty, if you can set up the
environment for it (cake with DOSBOX) so why not release it?

At least it will be a caution on the horrors of Olde Code.

I'm actually tempted to refactor it but we have other things
to do... this may be a good useless hobby, it it is still
useless and being useful and working at the highest level
possible subjectively is how all human beings, and the human
spirit of prosperity, advanced in the first place.

So I won't commit one way or the other, just continue to leave
this as a low priority.

Fun fact: this was supposed to be a demo for a programming job,
but I got the job without showing it and never needed to write
a line of C there. Instead, they became a JAVA shop, and then,
were elided from history despite all the excellent work they
did. But, that's another story, and a sad cautionary tale
not to take VCs too seriously.