Uploaded more recent builds
It might actually be possible to run this version now.

However, if it works, it is a happy surprise. It works on my box but the version I built is much more recent. 

Well, it cannot work without a sfchar.dat which matches the spec.

So ... this was pointless.

CWSDPMI.EXE might not be freely distributable, either. I have to put something about IPLefts in my Constitution.

or Konstitution.